move frontend specific util to its own file
avoids crashing importing frontend code in server components
