+++ lib/util.ts
@@ -1,8 +1,6 @@
 import { v4 as uuid } from 'uuid'
 import { ApiError } from '../models/ApiError'
 import { type Exercise } from '../models/AsyncSelectorOption/Exercise'
-import { enqueueSnackbar } from 'notistack'
-import { ERRORS } from './frontend/constants'
 
 /** Manually create a globally unique id across all tables. This should be used for ALL new records.
  We want to manually handle the IDs so that ID generation is not tied to the specific database being used,
@@ -76,20 +74,3 @@ export const removeUndefinedKeys = <T extends object>(obj: T) =>
       value !== undefined ? { ...prev, [key]: value } : prev,
     {}
   )
-
-export const enqueueError = (
-  e: unknown,
-  /** message to show if the error is a validation error */
-  validationMessage: string
-) => {
-  const originalMessage = e instanceof Error ? e.message : ''
-
-  enqueueSnackbar({
-    message:
-      originalMessage === ERRORS.validationFail
-        ? validationMessage
-        : ERRORS.retry,
-    severity: 'error',
-    persist: true,
-  })
-}
